+++ src/bootstrap/chromium/index.ts
@@ -1,12 +1,12 @@
 import { browser } from "webextension-polyfill-ts";
 
-import EventManager from "../../background/EventManager";
-import App from "../../background/App";
-import CertificateService from "../../background/certificate/CertificateService";
-import QualityProvider from "../../background/quality/providers/QualityProvider";
-import QualityService from "../../background/quality/QualityService";
-import ServerProvider from "../../background/certificate/providers/ServerProvider";
-import Configurator from "../../helpers/Configurator";
+import { App } from "../../background/App";
+import { CertificateService } from "../../background/certificate/CertificateService";
+import { ServerProvider } from "../../background/certificate/providers/ServerProvider";
+import { EventManager } from "../../background/EventManager";
+import { QualityProvider } from "../../background/quality/providers/QualityProvider";
+import { QualityService } from "../../background/quality/QualityService";
+import { Configurator } from "../../helpers/Configurator";
 
 const {
   browserAction: { setIcon, setBadgeText, setBadgeBackgroundColor },
@@ -14,7 +14,7 @@ const {
   storage,
   tabs: { onActivated },
   webNavigation: { onErrorOccurred },
-  webRequest: { onHeadersReceived },
+  webRequest: { onBeforeRequest, onHeadersReceived },
 } = browser;
 
 const certificateProvider = new ServerProvider();
@@ -29,6 +29,7 @@ configurator.init();
 const app = new App(certificateService, qualityService, configurator);
 
 const eventManager = new EventManager(
+  onBeforeRequest,
   onHeadersReceived,
   onErrorOccurred,
   onMessage,