+++ src/popup/Popup.tsx
@@ -2,14 +2,11 @@ import React, { Component } from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { Tabs, Runtime } from "webextension-polyfill-ts";
 
-import Certificate from "../types/certificate/Certificate";
-import ErrorMessage from "../types/errors/ErrorMessage";
+import { Certificate } from "../types/certificate/Certificate";
+import { ErrorMessage } from "../types/errors/ErrorMessage";
 import { Quality } from "../types/Quality";
-import CertificateView from "./pages/CertificateView";
-import Configuration from "./pages/Configuration";
-import Domains from "./pages/Domains";
-import Home from "./pages/Home";
-import NewQuality from "./pages/NewQuality";
+import { CertificateView } from "./pages/CertificateView";
+import { Home } from "./pages/Home";
 
 /**
  * Represents the required props for the Popup component
@@ -37,7 +34,7 @@ interface State {
  * Represents a browser action popup window
  * @noInheritDoc
  */
-export default class Popup extends Component<Props, State> {
+export class Popup extends Component<Props, State> {
   private sendMessage;
   private getTabs;
 
@@ -132,16 +129,7 @@ export default class Popup extends Component<Props, State> {
       <Router>
         <Switch>
           <Route path="/certificate">
-            <CertificateView />
-          </Route>
-          <Route path="/new-quality">
-            <NewQuality />
-          </Route>
-          <Route path="/domains">
-            <Domains />
-          </Route>
-          <Route path="/configuration">
-            <Configuration />
+            <CertificateView certificate={this.state.certificate} />
           </Route>
           <Route path="/">
             <Home