+++ src/popup/components/Navigation.tsx
@@ -1,11 +1,12 @@
 import React, { Component } from "react";
+import { NavLink } from "react-router-dom";
 import { Menu, Icon, Popup as Tooltip } from "semantic-ui-react";
 
 interface Props {
   openOptionsPage: () => Promise<void>;
 }
 
-export default class Navigation extends Component<Props> {
+export class Navigation extends Component<Props> {
   constructor(props: Props) {
     super(props);
     this.openOptionsPage = this.openOptionsPage.bind(this);
@@ -21,24 +22,8 @@ export default class Navigation extends Component<Props> {
         <Tooltip
           content="Certificate"
           trigger={
-            <Menu.Item disabled>
+            <Menu.Item as={NavLink} to="/certificate">
               <Icon name="clipboard outline" />
-            </Menu.Item>
-          }
-        />
-        <Tooltip
-          content="New Quality Entry"
-          trigger={
-            <Menu.Item disabled>
-              <Icon name="pencil" />
-            </Menu.Item>
-          }
-        />
-        <Tooltip
-          content="Managed Domains"
-          trigger={
-            <Menu.Item disabled>
-              <Icon name="list" />
             </Menu.Item>
           }
         />