+++ src/popup/pages/Home.tsx
@@ -2,19 +2,18 @@ import React, { Component } from "react";
 import {
   Container,
   Divider,
-  Header,
   Icon,
   Message,
   Popup as Tooltip,
   Rating,
-  Image,
   Placeholder,
 } from "semantic-ui-react";
 
-import Certificate from "../../types/certificate/Certificate";
-import ErrorMessage from "../../types/errors/ErrorMessage";
+import { Certificate } from "../../types/certificate/Certificate";
+import { ErrorMessage } from "../../types/errors/ErrorMessage";
 import { Quality, maxQuality } from "../../types/Quality";
-import Navigation from "../components/Navigation";
+import { Navigation } from "../components/Navigation";
+import { PageHeader } from "../components/PageHeader";
 
 interface Props {
   errorMessage: ErrorMessage | undefined;
@@ -22,7 +21,8 @@ interface Props {
   quality: Quality | undefined;
   openOptionsPage: () => Promise<void>;
 }
-export default class Home extends Component<Props> {
+
+export class Home extends Component<Props> {
   constructor(props: Props) {
     super(props);
   }
@@ -30,9 +30,7 @@ export default class Home extends Component<Props> {
   render(): JSX.Element {
     return (
       <div className="flex-container">
-        <Header as="h2" textAlign="center">
-          <Image src="../assets/logo.svg" size="tiny" /> Chexxo
-        </Header>
+        <PageHeader title="Chexxo" hasHomeButton={false} />
 
         <Divider section />