+++ src/types/errors/certificate/InvalidDomainError.ts
@@ -1,9 +1,9 @@
-import CertificateError from "./CertificateError";
+import { CertificateError } from "./CertificateError";
 
 /**
  * If a certificate is not valid for the requested domain this error gets thrown.
  */
-export default class InvalidDomainError extends CertificateError {
+export class InvalidDomainError extends CertificateError {
   readonly name: string = "Certificate invalid host exception";
   readonly code: number = 102;
   readonly message: string = "The certificate is not valid for this domain.";