+++ src/types/errors/certificate/RevokedError.ts
@@ -1,6 +1,6 @@
-import CertificateError from "./CertificateError";
+import { CertificateError } from "./CertificateError";
 
-export default class RevokedError extends CertificateError {
+export class RevokedError extends CertificateError {
   readonly name: string = "Certificate revoked exception";
   readonly code: number = 105;
   readonly message: string = "The Certificate was revoked.";