+++ src/types/errors/certificate/SelfSignedError.ts
@@ -1,9 +1,9 @@
-import CertificateError from "./CertificateError";
+import { CertificateError } from "./CertificateError";
 
 /**
  * If a certificate has not been signed by a trusted entity this error gets thrown.
  */
-export default class SelfSignedError extends CertificateError {
+export class SelfSignedError extends CertificateError {
   readonly name: string = "Certificate self signed exception";
   readonly code: number = 103;
   readonly message: string =