+++ src/types/errors/certificate/UnknownError.ts
@@ -1,6 +1,6 @@
-import CertificateError from "./CertificateError";
+import { CertificateError } from "./CertificateError";
 
-export default class UnknownError extends CertificateError {
+export class UnknownError extends CertificateError {
   readonly name: string = "Unknown certificate exception";
   readonly code: number = 106;
   readonly message: string =